Updated scripts for Raspberry Pi

+++ base-vm/scripts/main-node/look-and-feel.sh
@@ -1,25 +1,32 @@
 #!/bin/bash -x
 set -euo pipefail
 
-# Install KDE-Plasma GUI
-sudo DEBIAN_FRONTEND=noninteractive apt install -y sddm kde-plasma-desktop
+sudo DEBIAN_FRONTEND=noninteractive apt install -y sddm
 
-# Copy files needed for KX.AS.CODE look and file to relevant places
-sudo mkdir -p /usr/share/logos/
-sudo cp ${INSTALLATION_WORKSPACE}/theme/logos/* /usr/share/logos/
+# Only switch to KDE Plasma if not running on Raspberry Pi
+if [[ -z $(which raspinfo) ]]; then
 
-# Copy files needed for KX.AS.CODE look and file to relevant places
-sudo mkdir -p /usr/share/backgrounds/
-sudo cp ${INSTALLATION_WORKSPACE}/theme/backgrounds/* /usr/share/backgrounds/
-sudo rm -rf /usr/share/wallpapers/*
+  # Install KDE-Plasma GUI
+  sudo DEBIAN_FRONTEND=noninteractive apt install -y kde-plasma-desktop
 
-# Change SDDM Login Screen
-sudo apt-get install -y qtbase5-dev qtchooser qt5-qmake qtbase5-dev-tools
-sudo apt install -y \
-    qml-module-qtquick-controls \
-    qml-module-qtquick-extras \
-    qml-module-qtquick-layouts \
-    qml-module-qtgraphicaleffects
+  # Copy files needed for KX.AS.CODE look and file to relevant places
+  sudo mkdir -p /usr/share/logos/
+  sudo cp ${INSTALLATION_WORKSPACE}/theme/logos/* /usr/share/logos/
+
+  # Copy files needed for KX.AS.CODE look and file to relevant places
+  sudo mkdir -p /usr/share/backgrounds/
+  sudo cp ${INSTALLATION_WORKSPACE}/theme/backgrounds/* /usr/share/backgrounds/
+  sudo rm -rf /usr/share/wallpapers/*
+
+  # Change SDDM Login Screen
+  sudo apt-get install -y qtbase5-dev qtchooser qt5-qmake qtbase5-dev-tools
+  sudo apt install -y \
+      qml-module-qtquick-controls \
+      qml-module-qtquick-extras \
+      qml-module-qtquick-layouts \
+      qml-module-qtgraphicaleffects
+
+fi
 
 # Change SDDM Login Screen
 sudo cp -r ${INSTALLATION_WORKSPACE}/theme/sddm/chili-0.1.5 /usr/share/sddm/themes
